requirements.txt
streamlit
openai
streamlit_carousel
python-dotenv